refactor: extract rx_cli module, finalize WAV on start_rx failure
Review feedback (CodeRabbit + Codacy Lizard NLOC):
- The CLI layer (Args, usage, Config, build_config and its new
  resolve_freq / resolve_capture_path / validate_ranges helpers,
  print_verbose) moves to airspy_tools::rx_cli with its tests,
  halving the binary's file NLOC and putting the whole argument
  surface in the tested library.
- open_configured_device and run() shed their long tails into
  print_samplerate / print_device_serial / write_wav_placeholder /
  install_signal_handler helpers; every function is under the
  50-NLOC limit.
- New abort_stream cleanup: both start_rx and set_freq failures now
  stop the stream, drop the device, and finalize the output so a -w
  capture keeps a valid WAV header.
